requirements.txt
pandas
praw
python-dotenv
boto3
plotly
dash
